+++ src/custom-code-helpers/string-array/templates/string-array-rotate-function/StringArrayRotateFunctionTemplate.ts
@@ -3,14 +3,23 @@
  */
 export function StringArrayRotateFunctionTemplate (): string {
     return `
-        (function (array, {timesName}) {
-            const {whileFunctionName} = function (times) {
-                while (--times) {
+        (function (array, comparisonValue) {
+            array[-1] = true;
+
+            while (true) {
+                try {
+                    const expression = {comparisonExpressionCode};
+                                            
+                    if (expression === comparisonValue) {
+                        array[-1] = false;
+                        break;
+                    } else {
+                        array['push'](array['shift']());
+                    }
+                } catch (e) {
                     array['push'](array['shift']());
                 }
-            };
-            
-            {code}
-        })({stringArrayName}, {stringArrayRotationAmount});
+            }
+        })({stringArrayName}, {comparisonValue});
     `;
 }