+++ src/node-transformers/converting-transformers/NumberToNumericalExpressionTransformer.ts
@@ -15,7 +15,7 @@ import { NodeTransformationStage } from '../../enums/node-transformers/NodeTrans
 import { AbstractNodeTransformer } from '../AbstractNodeTransformer';
 import { NodeGuards } from '../../node/NodeGuards';
 import { NodeFactory } from '../../node/NodeFactory';
-import { NumberUtils } from '../../utils/NumberUtils';
+import { NumericalExpressionDataToNodeConverter } from '../../node/NumericalExpressionDataToNodeConverter';
 
 /**
  * replaces:
@@ -87,89 +87,18 @@ export class NumberToNumericalExpressionTransformer extends AbstractNodeTransfor
 
         const numberNumericalExpressionData: TNumberNumericalExpressionData = this.numberNumericalExpressionAnalyzer.analyze(literalNode.value);
 
-        return this.convertNumericalExpressionDataToNode(numberNumericalExpressionData);
-    }
+        return NumericalExpressionDataToNodeConverter.convert(
+            numberNumericalExpressionData,
+            (number: number, isPositiveNumber: boolean) => {
+                const numberLiteralNode: ESTree.Literal = NodeFactory.literalNode(number);
 
-    /**
-     * @param {TNumberNumericalExpressionData} numberNumericalExpressionData
-     * @param {ESTree.BinaryOperator} operator
-     * @returns {ESTree.BinaryExpression | ESTree.Literal | ESTree.UnaryExpression}
-     */
-    private convertNumericalExpressionDataToNode (
-        numberNumericalExpressionData: TNumberNumericalExpressionData,
-        operator: ESTree.BinaryOperator = '+'
-    ): ESTree.BinaryExpression | ESTree.Literal | ESTree.UnaryExpression {
-        const numberNumericalExpressionDataLength: number = numberNumericalExpressionData.length;
-
-        const leftParts: TNumberNumericalExpressionData = numberNumericalExpressionDataLength > 1
-            ? numberNumericalExpressionData.slice(0, numberNumericalExpressionDataLength - 1)
-            : [numberNumericalExpressionData[0]];
-        const rightParts: TNumberNumericalExpressionData = numberNumericalExpressionDataLength > 1
-            ? numberNumericalExpressionData.slice(-1)
-            : [];
-
-        // trailing iterations
-        if (rightParts.length) {
-            return this.convertPartsToBinaryExpression(operator, leftParts, rightParts);
-        }
-
-        const firstLeftPartOrNumber: number | number[] | null = leftParts[0] ?? null;
-
-        // last iteration when only single left part is left
-        return Array.isArray(firstLeftPartOrNumber)
-            ? this.convertNumericalExpressionDataToNode(firstLeftPartOrNumber, '*')
-            : this.convertPartOrNumberToLiteralNode(firstLeftPartOrNumber);
-    }
-
-    /**
-     * @param {ESTree.BinaryOperator} operator
-     * @param {TNumberNumericalExpressionData} leftParts
-     * @param {TNumberNumericalExpressionData} rightParts
-     * @returns {ESTree.BinaryExpression}
-     */
-    private convertPartsToBinaryExpression (
-        operator: ESTree.BinaryOperator,
-        leftParts: TNumberNumericalExpressionData,
-        rightParts: TNumberNumericalExpressionData
-    ): ESTree.BinaryExpression {
-        const rightPartOrNumber: number | number[] = rightParts[0];
-
-        if (Array.isArray(rightPartOrNumber)) {
-            // right part is array with multiply numbers
-            return NodeFactory.binaryExpressionNode(
-                operator,
-                this.convertNumericalExpressionDataToNode(leftParts),
-                this.convertNumericalExpressionDataToNode(rightPartOrNumber, '*')
-            );
-        } else {
-            // right part is number
-            return NodeFactory.binaryExpressionNode(
-                operator,
-                this.convertNumericalExpressionDataToNode(leftParts),
-                this.convertPartOrNumberToLiteralNode(rightPartOrNumber)
-            );
-        }
-    }
-
-    /**
-     * @param {number | number[]} partOrNumber
-     * @returns {ESTree.Literal}
-     */
-    private convertPartOrNumberToLiteralNode (
-        partOrNumber: number | number[]
-    ): ESTree.Literal | ESTree.UnaryExpression {
-        const number: number = Array.isArray(partOrNumber)
-            ? partOrNumber[0]
-            : partOrNumber;
-        const isPositiveNumber: boolean = NumberUtils.isPositive(number);
-
-        const literalNode: ESTree.Literal = NodeFactory.literalNode(Math.abs(number));
-
-        return isPositiveNumber
-            ? literalNode
-            : NodeFactory.unaryExpressionNode(
-                '-',
-                literalNode
-            );
+                return isPositiveNumber
+                    ? numberLiteralNode
+                    : NodeFactory.unaryExpressionNode(
+                        '-',
+                        numberLiteralNode
+                    );
+            }
+        );
     }
 }